refactor: renamed media to display
