+++ docs/src/pages/display/Icon.md
