+++ docs/src/pages/display/Image.md
