+++ docs/src/pages/display/Progress.md
@@ -1,0 +1,9 @@
+## Progress
+
+The **Progress** component represents the current status of an ongoing task or action.
+
+```jsx
+render(() => (
+	<Progress value={25} max={100} />
+))
+```
