+++ lib/components/display/IImageProps.tsx
