+++ lib/components/display/Image.tsx
