+++ test/components/display/Icon.test.tsx
