+++ test/components/display/Image.test.tsx
